+++ src/tileray.h
@@ -46,7 +46,7 @@ class tileray
         int dir() const;      // return direction of ray (degrees)
         int dir4() const;     // return 4-sided direction (0 = east, 1 = south, 2 = west, 3 = north)
         int dir8() const;     // return 8-sided direction (0 = east, 1 = southeast, 2 = south ...)
-        long dir_symbol( long sym )
+        int dir_symbol( int sym )
         const; // convert certain symbols from north-facing variant into current dir facing
         int ortho_dx( int od ) const; // return dx for point at "od" distance in orthogonal direction
         int ortho_dy( int od ) const; // return dy for point at "od" distance in orthogonal direction
